% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_kernels.R
\name{ClusteringAllocator}
\alias{ClusteringAllocator}
\title{Allocates and initializes input to the clustering for machine learning
microbenchmarks}
\usage{
ClusteringAllocator(benchmarkParameters)
}
\arguments{
\item{benchmarkParameters}{an object of type
\code{\link{ClusteringMicrobenchmark}} specifying various parameters
needed to generate input for the clustering microbenchmarks.}
}
\value{
a list containing the data objects to be input to the
  clustering microbenchmark
}
\description{
\code{ClusteringAllocator} allocates and initializes the data sets that are
input to the clustering microbenchmarks for the purposes of conducting a
single performance trial with one of the clustering microbenchmark functions.
}
\seealso{
\code{\link{PamClusteringMicrobenchmark}} \code{\link{ClaraClusteringMicrobenchmark}}
}
