% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_utils.R
\name{ComputeAverageTime}
\alias{ComputeAverageTime}
\title{Computes the average of a vector of performance trial times}
\usage{
ComputeAverageTime(numberOfSuccessfulTrials, times)
}
\arguments{
\item{numberOfSuccessfulTrials}{the number of successful performance trials
to be averaged over}

\item{times}{a vector of wall clock times for the performance trials}
}
\description{
\code{ComputeAverageTime} computes the average of a vector of performance
  trial times.  The average is computed only over the first
  \code{numberOfSuccessfulTrials} elements of the \code{times} vector.
}
