% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_utils.R
\name{ComputeStandardDeviation}
\alias{ComputeStandardDeviation}
\title{Computes the standard deviation of a vector of performance trial times}
\usage{
ComputeStandardDeviation(numberOfSuccessfulTrials, times)
}
\arguments{
\item{numberOfSuccessfulTrials}{the number of successful performance trials
over which the standard deviation will be computed}

\item{times}{a vector of wall clock times for the performance trials}
}
\description{
\code{ComputeStandardDeviation} computes the standard deviation of a vector
  of performance trial times.  The standard deviation is computed only over
  the first \code{numberOfSuccessfulTrials} elements of the \code{times}
  vector.
}
