% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_benchmark.R
\name{GetClusteringExampleMicrobenchmarks}
\alias{GetClusteringExampleMicrobenchmarks}
\title{Initializes the list of example clustering microbenchmarks}
\usage{
GetClusteringExampleMicrobenchmarks()
}
\value{
a list of \code{ClusteringMicrobenchmark} objects defining the
  microbenchmarks to be executed.  Microbenchmarks for the \code{pam}
  and \code{clara} functions from the \code{cluster} package are
  provided.
}
\description{
\code{GetClusteringExampleMicrobenchmarks} defines the example clustering
microbenchmarks to be executed by the
\code{\link{RunMachineLearningBenchmark}} function.  The example
are chosen so that they can run in a few minutes or less.
\enumerate{
   \item{pam_cluster_3_7_2500}{N=3, seven clusters with 2500 vectors per
     cluster}
   \item{clara_cluster_64_33_1213}{N=64, 33 clusters with 1213 vectors per
     cluster}
}
See the documentation for the
\code{\link{ClusteringMicrobenchmark}} class for more details.
}
\seealso{
Other machine learning default microbenchmarks: \code{\link{GetClusteringDefaultMicrobenchmarks}}
}
