% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_utils.R
\name{GetConfigurableEnvParameter}
\alias{GetConfigurableEnvParameter}
\title{Retrieves the value of an environment variable referenced by another
environment variable}
\usage{
GetConfigurableEnvParameter(configurableVariable)
}
\arguments{
\item{configurableVariable}{a string parameter containing the name
of an environment variable which itself references another 
environment variable}
}
\value{
the value of the environment variable referenced by the
  environment variable specified in the \code{configurableVariable}
  parameter
}
\description{
\code{GetConfigurableEnvParameters} returns the value of the environment
variable referenced by the argument \code{configurableVariable} which
is also an environment variable
}
\details{
This function takes the argument \code{configurableVariable} which
contains the name of an environment variable whose value is an
environment variable referencing a value to be returned by this
function.
}
