% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense_matrix_benchmark.R
\name{GetDenseMatrixDefaultMicrobenchmarks}
\alias{GetDenseMatrixDefaultMicrobenchmarks}
\title{Initializes the list of default dense matrix microbenchmarks}
\usage{
GetDenseMatrixDefaultMicrobenchmarks()
}
\value{
a list of \code{DenseMatrixMicrobenchmark} objects defining the
  microbenchmarks to be executed.  The microbenchmarks appear in the order
  listed in the function description and are assigned the following names:
  cholesky, crossprod, determinant, eigen, solve, lsfit, deformtrans,
  transpose, matmat, matvec, qr, and svd.
}
\description{
\code{GetDenseMatrixDefaultMicrobenchmarks} defines the default dense
matrix microbenchmarks to be executed by the
\code{\link{RunDenseMatrixBenchmark}} function.  The current microbenchmarks
are Cholesky factorization, matrix cross product, matrix determinant,
eigendecomposition, linear solve with multiple right hand sides, least
squares fit, matrix deformation and transpose, matrix-matrix multiplication,
matrix-vector multiplication, QR decomposition, and singular value
decomposition.  See the documentation for the
\code{\link{DenseMatrixMicrobenchmark}} class for more details.
}
\seealso{
\code{\link{DenseMatrixMicrobenchmark}}
}
