% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense_matrix_benchmark.R
\name{GetDenseMatrixExampleMicrobenchmarks}
\alias{GetDenseMatrixExampleMicrobenchmarks}
\title{Initializes the list of example dense matrix microbenchmarks}
\usage{
GetDenseMatrixExampleMicrobenchmarks()
}
\value{
a list of \code{DenseMatrixMicrobenchmark} objects defining the
  microbenchmarks to be executed.  Microbenchmarks for Cholesky factorization
  and matrix cross product are provided.
}
\description{
\code{GetDenseMatrixExampleMicrobenchmarks} defines example dense
matrix microbenchmarks to be executed by the examples section
of the \code{\link{RunDenseMatrixBenchmark}} function.  The examples
are chosen so that they can run in a few minutes or less.
}
