% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_utils.R
\name{GetNumberOfThreads}
\alias{GetNumberOfThreads}
\title{Retrieves the number of threads from the environment}
\usage{
GetNumberOfThreads()
}
\value{
the number of threads retrieved from the environment
}
\description{
\code{GetNumberOfThreads} retrieves from the environment the number of
threads kernels are intended to be executed with
}
\details{
This function retrieves the number of threads kernels are intended to be
microbenchmarked with.  The number of threads is assumed to be stored
in an environment variable which this function retrieves.
}
