% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix_benchmark.R
\name{GetSparseCholeskyExampleMicrobenchmarks}
\alias{GetSparseCholeskyExampleMicrobenchmarks}
\title{Initializes the list of example sparse Cholesky factorization
microbenchmarks}
\usage{
GetSparseCholeskyExampleMicrobenchmarks()
}
\value{
a list of \code{SparseMatrixMicrobenchmark} objects defining the
  microbenchmarks to be executed.  The microbenchmark for the Cholesky
  factorization of the ct20stif matrix.
}
\description{
\code{GetSparseCholeskyExampleMicrobenchmarks} defines the example sparse
Cholesky factorization microbenchmarks to be executed by the
\code{\link{RunSparseMatrixBenchmark}} function.  The current sparse
Cholesky factorization microbenchmarks cover the following matrices:
\enumerate{
  \item cholesky_ct20stif -- Boeing structural matrix with 2600295 nonzeros
}
See the documentation for the
\code{\link{SparseMatrixMicrobenchmark}} class for more details.
}
\seealso{
Other sparse matrix default microbenchmarks: \code{\link{GetSparseCholeskyDefaultMicrobenchmarks}},
  \code{\link{GetSparseLuDefaultMicrobenchmarks}},
  \code{\link{GetSparseMatrixVectorDefaultMicrobenchmarks}},
  \code{\link{GetSparseMatrixVectorExampleMicrobenchmarks}},
  \code{\link{GetSparseQrDefaultMicrobenchmarks}}
}
