% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix_kernels.R
\name{SparseLuAllocator}
\alias{SparseLuAllocator}
\title{Allocates and initializes input to the LU factorization sparse matrix kernel
microbenchmarks}
\usage{
SparseLuAllocator(benchmarkParameters, index)
}
\arguments{
\item{benchmarkParameters}{an object of type
\code{\link{SparseMatrixMicrobenchmark}} specifying various parameters
needed to generate input for the sparse matrix kernel.}

\item{index}{an integer index indicating the dimensions of the matrix or
vector data to be generated as input for the sparse matrix kernel.}
}
\description{
\code{SparseLuAllocator} allocates and initializes the sparse matrix that is
input to the sparse matrix kernel for the purposes of conducting a single
performance trial with the \code{SparseCholeskyMicrobenchmark} function.  The
matrix is populated and returned in the \code{kernelParameters} list.
}
