% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix_kernels.R
\name{SparseMatrixVectorAllocator}
\alias{SparseMatrixVectorAllocator}
\title{Allocates and initializes input to the matrix-vector multiplication sparse
matrix kernel microbenchmarks}
\usage{
SparseMatrixVectorAllocator(benchmarkParameters, index)
}
\arguments{
\item{benchmarkParameters}{an object of type
\code{\link{SparseMatrixMicrobenchmark}} specifying various parameters
needed to generate input for the sparse matrix kernel.}

\item{index}{an integer index indicating the dimensions of the matrix or
vector data to be generated as input for the sparse matrix kernel.}
}
\value{
a list containing the matrices or vectors to be input for the
  sparse matrix kernel for which a single performance trial is to be
  conducted.
}
\description{
\code{SparseMatrixVectorAllocator} allocates and initializes the sparse
matrix and vector that are inputs to the sparse matrix kernel for the
purposes of conducting a single performance trial with the
\code{SparseMatrixVectorMicrobenchmark} function.  The matrix and vector
are populated and returned in the \code{kernelParameters} list.
}
